requirements.txt
openai~=0.27.2
tenacity~=8.2.2
tiktoken
numpy~=1.24.2
pandas~=1.5.3
matplotlib~=3.7.1
plotly~=5.13.1
scipy~=1.10.1
scikit-learn~=1.2.2
pydantic~=1.10.7
pyyaml~=5.1
mock~=5.0.1